AIG_Llama2 - 26-12/requirements.txt
fastapi
uvicorn
requests==2.28.2
accelerate
tika==1.24
aspose-words
beautifulsoup4
auto_gptq
transformers